[ALPS03194785] M4U:fix m4u code defect issue
[Detail] m4u have defect logically dead code,
out of bounds, extra high-orders bits issue

[Solution]
1.fix m4u_hw.c out of bounds and logically dead code
2.fix m4u_pgtable.c extra high orders bits

MTK-Commit-Id: 1009607359283a78200ece0fde43d730cdc77b84

CR-Id: ALPS03194785
Feature: others

+++ drivers/misc/mediatek/m4u/2.0/m4u_pgtable.c
@@ -141,7 +141,7 @@ int m4u_get_pte_info(m4u_domain_t *domain, unsigned int mva, m4u_pte_info_t *pte
 {
 	imu_pgd_t *pgd;
 	imu_pte_t *pte = NULL;
-	unsigned int pa = 0;
+	unsigned long pa = 0;
 	unsigned int size;
 	int valid = 1;
